work in progress

+++ powershell/Delete-CloudflareDnsRecord.ps1
@@ -6,15 +6,33 @@
     [string]$ZoneId,
 
     [Parameter(Mandatory = $true)]
-    [string]$RecordId
+    [string]$DnsName
 
 )
 
 $ErrorActionPreference = "Stop"
 
-$url = "https://api.cloudflare.com/client/v4/zones/$ZoneId/dns_records/$RecordId"
+$existingDnsRecords = $( .\Get-CloudflareDnsRecords.ps1 -ApiToken $env:CLOUDFLARE_API_KEY -ZoneId "$zoneId" )
 
-$response = curl $url `
+Write-Host "Existing DNS records count: $( $existingDnsRecords.Count )"
+
+$dnsRecordExists = $existingDnsRecords.ContainsKey($DnsName)
+
+Write-Host "DNS name $DnsName exists: $dnsRecordExists"
+
+if ($dnsRecordExists -eq $False)
+{
+    Write-Host "DNS record $DnsName does not exist. Skipping deletion ..." -ForegroundColor Yellow
+    exit 0
+}
+
+Write-Host "DNS record $DnsName exists. Starting deletion ..." -ForegroundColor Green
+
+$recordId = $existingDnsRecords[$DnsName]
+
+$url = "https://api.cloudflare.com/client/v4/zones/$ZoneId/dns_records/$recordId"
+
+$response = curl -s -S $url `
     -X DELETE `
     -H "Authorization: Bearer $ApiToken" `
     -H "Content-Type: application/json"
@@ -23,13 +41,13 @@ $responseJson = $response | ConvertFrom-Json
 
 if ($responseJson.success -eq $true)
 {
-    Write-Host "DNS record $RecordId deleted successfully."
+    Write-Host "DNS record $DnsName ID: $recordId deleted successfully." -ForegroundColor Green
     Write-Host "Response: $responseJson"
     return $responseJson.result.id
 }
 else
 {
-    Write-Host "Failed to delete DNS record $RecordId."
+    Write-Host "Failed to delete DNS record $recordId." -ForegroundColor Red
     Write-Host "Response: $( $response )"
     exit 1
 }
